pyproject.toml
[tool.poetry]
name = "sp_ask_list_of_concurrent_chats"
version = "0.1.2"
description = "Concurrent chats on LibraryH3lp"
authors = ["Guinsly Mondesir <guinslym@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
lh3api = "^0.2.0"
numpy = "^1.18.2"
pandas = "^1.0.3"
openpyxl = "^3.0.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
